docker-demo/backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["haruiz <henryruiz22@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.90.1"
uvicorn = "^0.20.0"
fastapi-utils = "^0.2.1"
pydantic = "^1.10.4"
python-multipart = "^0.0.5"
python-decouple = "^3.7"
requests = "^2.28.2"
pillow = "^9.4.0"
#psycopg2-binary = "^2.9.5"
google-cloud-storage = "^2.7.0"


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
